requirements.txt
langgraph
langchain
langchain-openai
yfinance
pandas
pandas-ta
matplotlib
mplfinance
requests
python-dotenv
xlrd
tqdm
